doc-app/backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
sqlalchemy
psycopg2-binary
alembic
requests
pydantic
aiohttp
beautifulsoup4
lxml
pyjwt
python-multipart